+++ packages/cli-plugin-dev/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/midwayjs/cli/compare/v1.3.12-beta.3...v1.3.12) (2022-08-25)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-dev
+
+
+
+
+
+## [1.3.12-beta.3](https://github.com/midwayjs/cli/compare/v1.3.12-beta.2...v1.3.12-beta.3) (2022-08-22)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-dev
+
+
+
+
+
+## [1.3.12-beta.2](https://github.com/midwayjs/cli/compare/v1.3.12-beta.1...v1.3.12-beta.2) (2022-08-07)
+
+**Note:** Version bump only for package @midwayjs/cli-plugin-dev
+
+
+
+
+
+## [1.3.12-beta.1](https://github.com/midwayjs/cli/compare/v1.3.11...v1.3.12-beta.1) (2022-08-05)
+
+
+### Bug Fixes
+
+* dev handle error ([0bc3020](https://github.com/midwayjs/cli/commit/0bc3020f44e54a3a4966d33bf5c1851a404d6d08))
+
+
+
+
+
 ## [1.3.11](https://github.com/midwayjs/cli/compare/v1.3.9...v1.3.11) (2022-08-04)
 
 **Note:** Version bump only for package @midwayjs/cli-plugin-dev